requirements.txt
streamlit>=1.33
pandas>=2.1
openpyxl>=3.1
gTTS>=2.5
requests>=2.31
python-dotenv>=1.0
google-generativeai>=0.7
